Added Paging, now boots into LM
+ Paging
* Activate LM and jump into Kernel

+++ src/loader/include/types.h
@@ -22,7 +22,7 @@
 typedef unsigned char byte;
 typedef unsigned short word;
 typedef unsigned int dword;
-typedef unsigned long qword;
+typedef unsigned long long qword;
 
 typedef signed char sbyte;
 typedef signed short sword;
